Fix for 3618 again. Had one of those "Isn't there a term missing in that equation?" moments while debugging. z7s should load properly now. (This will be more important in the future when we have z6 support, since z6 addresses memory in the same way as z7. z7 in itself is rather pointless.)
custard.z7, the only z7 I can find, sometimes asks for the user's name multiple times, and sometimes inputs it at the top of the screen. Otherwise it works just like it does under Frotz. (The printed text under Frotz indicates that it's special-casing when it detects Frotz, so goodness knows what's supposed to happen.)